+++ lib/legion/extensions/governance.rb
@@ -4,6 +4,8 @@ require 'legion/extensions/governance/version'
 require 'legion/extensions/governance/helpers/layers'
 require 'legion/extensions/governance/helpers/proposal'
 require 'legion/extensions/governance/runners/governance'
+require 'legion/extensions/governance/runners/shadow_ai'
+require 'legion/extensions/governance/actors/shadow_ai_scan'
 
 module Legion
   module Extensions