+++ lib/legion/extensions/governance/actors/shadow_ai_scan.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Legion
+  module Extensions
+    module Governance
+      module Actors
+        class ShadowAiScan < Legion::Extensions::Actors::Every
+          def runner_class = Runners::ShadowAi
+          def runner_function = 'full_scan'
+          def time = 86_400
+        end
+      end
+    end
+  end
+end